Restuctured System level initializations into separate file.
