+++ SYSTEM.Initialize.h
@@ -1,0 +1,11 @@
+// SYSTEM.Initialize.h
+//////////////////////////////////////////////////////////////////////////
+
+#pragma once
+
+#include "Application.h"
+
+namespace INITIALIZE {
+	int WindowFailed();
+	int GlewFailed();
+}
