+++ cg_programming_windows.vcxproj.filters
@@ -48,6 +48,9 @@
     <ClCompile Include="SYSTEM.Mesh.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="SYSTEM.Initialize.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Application.h">
@@ -81,6 +84,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="SYSTEM.Mesh.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="SYSTEM.Initialize.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
